Remove CI/CD workflows, fix 400+ clippy warnings
- Removed .github/workflows/ci.yml and publish.yml (spamming, not functional)
- Fixed 400+ clippy warnings across 25 files:
  - 132 format string simplifications (inline variables)
  - 27 useless String conversions
  - 30 unused variables (prefixed with _)
  - Numerous dead assignments, mutable vars, and redundant closures
  - Reduced from 443 to 8 warnings (remaining are pre-existing unreachable patterns)

+++ crates/lightning-core/src/optimizer/cardinality_estimator.rs
@@ -61,8 +61,8 @@ impl CardinalityEstimator {
                     (child_card as f64).sqrt() as u64
                 }
             }
-            LogicalOperator::Limit(_, limit) => *limit as u64,
-            LogicalOperator::Skip(child, skip) => self.estimate(child).saturating_sub(*skip as u64),
+            LogicalOperator::Limit(_, limit) => *limit,
+            LogicalOperator::Skip(child, skip) => self.estimate(child).saturating_sub(*skip),
             LogicalOperator::Union(left, right, _) => self.estimate(left) + self.estimate(right),
             _ => 1000,
         }
